dx11: better clipping algo. modvol cull mode. win: exception code
dx11: use faster clipping algorithm for modifier volumes
dx11: modifier volumes need cull mode. Fixes Daytona shadows.
win: the fault handler was treating some software exceptions as fatal (msvc)
uwp: remove async hacks and disable MS task wait exception. Fixes error reporting when loading a game.

+++ core/rend/dx11/dx11_renderstate.h
@@ -94,7 +94,7 @@ public:
 
 	ComPtr<ID3D11DepthStencilState> getState(bool depth, bool depthWrite, int depthFunc, bool stencil)
 	{
-		int hash = (depthFunc << 3) | depth | (depthWrite << 1) | (stencil << 2);
+		int hash = (depthFunc << 3) | (int)depth | ((int)depthWrite << 1) | ((int)stencil << 2);
 		auto& state = states[hash];
 		if (!state)
 		{
@@ -186,7 +186,7 @@ class BlendStates
 public:
 	ComPtr<ID3D11BlendState> getState(bool enable, int srcBlend = 0, int destBlend = 0, bool disableWrite = false)
 	{
-		int hash = enable | (srcBlend << 1) | (destBlend << 5) | (disableWrite << 9);
+		int hash = (int)enable | (srcBlend << 1) | (destBlend << 5) | ((int)disableWrite << 9);
 		auto& state = states[hash];
 		if (!state)
 		{
